+++ app/app.js
@@ -10,6 +10,10 @@ var api = require('./routes/api');
 var db = require('./db');
 var app = express();
 
+// Viwe Engine
+app.set('views', path.join(__dirname, 'views'));
+app.set('view engine', 'pug');
+
 app.use(favicon(path.join(__dirname, '../public', 'favicon.ico')));
 app.use(logger('dev'));
 app.use(bodyParser.json());